pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "home-optimizer"
version = "0.10"
description = "Slimme warmtepompregeling met MPC en automatische kalibratie, gericht op maximale zelfconsumptie van zonne-energie en minimale energiekosten."
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "gerjanvdbosch" },
]
dependencies = [
  "plotly>=5.20",
  "httpx>=0.27",
  "pandas>=2.2",
  "numpy>=1.26",
  "fastapi>=0.110",
  "uvicorn>=0.29",
  "pydantic>=2.7",
  "SQLAlchemy>=2.0",
  "APScheduler>=3.10",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
  "black>=24.0",
  "pyflakes>=3.2",
  "pyright>=1.1,<2.0",
  "pytest>=8.0",
  "ruff>=0.4",
  "import-linter>=2.0",
]

[project.scripts]
home-optimizer-addon = "home_optimizer.entrypoints.addon:main"
home-optimizer-local = "home_optimizer.entrypoints.local:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
home_optimizer = ["templates/*.html"]
"home_optimizer.web" = ["templates/*.html", "static/*.css", "static/*.js"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests", "examples"]

[tool.ruff.lint]
select = ["E", "F", "I", "B"]

[tool.ruff.lint.per-file-ignores]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.pyright]
include = ["src", "tests"]
exclude = ["**/__pycache__"]
extraPaths = ["src"]
venvPath = "."
venv = ".venv"
pythonVersion = "3.11"
typeCheckingMode = "standard"
reportMissingTypeStubs = false

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.importlinter]
root_package = "home_optimizer"

[[tool.importlinter.contracts]]
name = "Clean architecture layers"
type = "layers"
layers = [
  "home_optimizer.entrypoints",
  "home_optimizer.web",
  "home_optimizer.app",
  "home_optimizer.infrastructure",
  "home_optimizer.features",
  "home_optimizer.domain",
]
